Added the Menu
i have added the menu interface where we can click on the options so it will start the game and so on.
i have added a pause button as well while playing the game if the letter P is pressed the game can be paused.
i have edited the main gameloop here as well.

Co-Authored-By: SapnaJaved <188207494+SapnaJaved@users.noreply.github.com>